chore: partially clean-up remaining examples
