+++ Examples/DemoApp/DemoApp/Assets.xcassets/AccentColor.colorset/Contents.json
