+++ Examples/DemoApp/DemoApp/Assets.xcassets/AppIcon.appiconset/Contents.json
