+++ Examples/DemoApp/DemoApp/Assets.xcassets/Contents.json
