+++ Examples/DemoApp/DemoApp/Info.plist
