+++ Examples/DemoApp/DemoAppTests/Tests/en.lproj/InfoPlist.strings
