+++ Examples/DemoApp/MUXSDKImaListener.xcworkspace/xcshareddata/WorkspaceSettings.xcsettings
