+++ Examples/MuxStatsGoogleIMAPluginCocoapodsExampleTVOS/MuxStatsGoogleIMAPluginCocoapodsExampleTVOS.xcworkspace/contents.xcworkspacedata
@@ -1,0 +1,10 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<Workspace
+   version = "1.0">
+   <FileRef
+      location = "group:MuxStatsGoogleIMAPluginCocoapodsExampleTVOS.xcodeproj">
+   </FileRef>
+   <FileRef
+      location = "group:Pods/Pods.xcodeproj">
+   </FileRef>
+</Workspace>
