+++ Examples/MuxStatsGoogleIMAPluginCocoapodsExampleTVOS/MuxStatsGoogleIMAPluginCocoapodsExampleTVOS/ProcessInfo+Data.swift
@@ -1,9 +1,6 @@
 //
 //  ProcessInfo+Data.swift
-//  MUXSDKIMATVOSExample
-//
-//  Created by AJ Barinov on 9/6/23.
-//  Copyright © 2023 Dylan Jhaveri. All rights reserved.
+//  MuxStatsGoogleIMAPluginCocoapodsExampleTVOS
 //
 
 import Foundation
