+++ Examples/MuxStatsGoogleIMAPluginCocoapodsExampleTVOS/MuxStatsGoogleIMAPluginCocoapodsExampleTVOSUITests/MuxStatsGoogleIMAPluginCocoapodsExampleTVOSUITests.swift
@@ -1,14 +1,13 @@
 //
-//  MUXSDKIMATVOSExampleUITests.swift
-//  MUXSDKIMATVOSExampleUITests
+//  MuxStatsGoogleIMAPluginCocoapodsExampleTVOSUITests.swift
+//  MuxStatsGoogleIMAPluginCocoapodsExampleTVOSUITests
 //
-//  Created by AJ Barinov on 9/1/23.
-//  Copyright © 2023 Dylan Jhaveri. All rights reserved.
+//  Created by AJ Barinov on 11/7/24.
 //
 
 import XCTest
 
-final class MUXSDKIMATVOSExampleUITests: XCTestCase {
+final class MuxStatsGoogleIMAPluginCocoapodsExampleTVOSUITests: XCTestCase {
 
     override func setUpWithError() throws {
         // Put setup code here. This method is called before the invocation of each test method in the class.
@@ -23,6 +22,7 @@ final class MUXSDKIMATVOSExampleUITests: XCTestCase {
         // Put teardown code here. This method is called after the invocation of each test method in the class.
     }
 
+    @MainActor
     func testExample() throws {
         // UI tests must launch the application that they test.
         let app = XCUIApplication()
@@ -31,6 +31,7 @@ final class MUXSDKIMATVOSExampleUITests: XCTestCase {
         // Use XCTAssert and related functions to verify your tests produce the correct results.
     }
 
+    @MainActor
     func testLaunchPerformance() throws {
         if #available(macOS 10.15, iOS 13.0, tvOS 13.0, watchOS 7.0, *) {
             // This measures how long it takes to launch your application.
