--- Examples/Podfile
@@ -1,26 +1,0 @@
-use_frameworks!
-
-target 'DemoApp' do
-  platform :ios, '12.0'
-  pod 'Mux-Stats-Google-IMA', :path => '../'
-end
-
-target 'MUXSDKIMATVOSExample' do
-  platform :tvos, '14.0'
-  pod 'Mux-Stats-Google-IMA', :path => '../'
-end
-
-target 'DemoAppUITests' do
-  platform :ios, '12.0'
-  pod 'Mux-Stats-Google-IMA', :path => '../'
-end
-
-post_install do |installer|
-    installer.generated_projects.each do |project|
-          project.targets.each do |target|
-              target.build_configurations.each do |config|
-                  config.build_settings['IPHONEOS_DEPLOYMENT_TARGET'] = '12.0'
-               end
-          end
-   end
-end
